+++ packages/acme-view/src/components/Hero/Wrap.js
@@ -1,7 +1,13 @@
 import React from 'react'
 import Hero from '.'
-import image from '../../../assets/orion-hero-2x.png'
 
-export default function() {
-  return <Hero imageSrc={image} />
+export default function({ image, subtitle, title }) {
+  return (
+    <Hero
+      imageSrc={image}
+      subtitle={subtitle}
+      title={title}
+      onSearch={() => {}}
+    />
+  )
 }