+++ packages/gatsby-plugin-orion-core/src/components/Footer/index.js
@@ -1,9 +1,10 @@
 import React from 'react'
 import T from 'prop-types'
 import { withStyles, Grid, Typography } from '@material-ui/core'
+import { Link } from '@reach/router'
 import PaddedContainer from '../PaddedContainer'
 
-function Footer({ classes, Img, socialIcons = [], logo }) {
+function Footer({ classes, Img, socialIcons = [], Logo }) {
   return (
     <div className={classes.footer}>
       <PaddedContainer>
@@ -23,8 +24,8 @@ function Footer({ classes, Img, socialIcons = [], logo }) {
           </Grid>
           <Grid container item xs={4} justify="flex-end">
             <Grid container item xs={12} justify="flex-end">
-              <Grid item component="a" href="/">
-                <img src={logo} alt="Logo" className="logo" />
+              <Grid item component={Link} to="/">
+                <Logo width="107" />
               </Grid>
             </Grid>
             <Grid
@@ -57,8 +58,8 @@ function Footer({ classes, Img, socialIcons = [], logo }) {
             <Typography
               display="inline"
               variant="body1"
-              component="a"
-              href="/terms"
+              component={Link}
+              to="/terms"
             >
               Terms of Use
             </Typography>
@@ -67,8 +68,8 @@ function Footer({ classes, Img, socialIcons = [], logo }) {
             <Typography
               display="inline"
               variant="body1"
-              component="a"
-              href="/privacy"
+              component={Link}
+              to="/privacy"
             >
               Privacy Statement
             </Typography>
@@ -85,13 +86,13 @@ Footer.propTypes = {
   classes: T.object,
   Img: T.elementType.isRequired,
   socialIcons: T.arrayOf(T.object),
-  logo: T.string,
+  Logo: T.element,
 }
 
 Footer.defaultProps = {
   classes: {},
   socialIcons: [],
-  logo: '',
+  Logo: undefined,
 }
 
 export default withStyles(styles, { withTheme: true })(Footer)